# docs/openapi/teamclu-api.v1.yaml
openapi: 3.1.0
info:
  title: TeamClu Cloud API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.teamclu.local
security:
  - bearerAuth: []
...
paths:
...
  /v1/apps:
    get:
      operationId: listApps
      summary: List the apps visible to the caller in a team
      tags: [Apps]
      parameters:
        - name: teamId
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Visible apps, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/App" }
        "400":
          $ref: "#/components/responses/Error"
    post:
      operationId: createApp
      summary: Create an app (provisions its 1:1 workspace)
      tags: [Apps]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId, name, type]
              properties:
                teamId: { type: string }
                name: { type: string }
                type:
                  $ref: "#/components/schemas/AppType"
                visibility: { type: string, enum: [personal, team], default: personal }
                gitRemoteUrl:
                  type: [string, "null"]
                  description: >-
                    Optional repository to import. When given, the local daemon
                    clones it into the app's checkout and writes no starter
                    template; `type` then only decides how the app deploys.
                    Must be an http(s), ssh or git:// address, or the scp-like
                    `git@host:owner/repo.git` form — anything else is rejected
                    with 400, since git reads some other forms as commands to
                    run rather than as addresses.
                  example: https://github.com/owner/repo.git
      responses:
        "201":
          description: |
            The created app. `provisionStatus` is `repo_created` once the
            the workspace exists, or `error` when provisioning failed —
            the row is created either way.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "400":
          $ref: "#/components/responses/Error"
  /v1/apps/{appId}:
    get:
      operationId: getApp
      summary: Get one app
      tags: [Apps]
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: The app.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateApp
      summary: Update an app (creator only)
      description: |
        Renames, changes visibility, or reports a lifecycle status the client
        owns. `provisionStatus` carries the repo/seed lifecycle the desktop
        drives via the local daemon; `fcStatus` carries the deploy lifecycle —
        the desktop reports `deploy_error` when the daemon build leg fails,
        since nothing server-side can observe that. Illegal transitions are
        rejected.
      tags: [Apps]
      parameters:
        - $ref: "#/components/parameters/AppId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                visibility: { type: string, enum: [personal, team] }
                provisionStatus: { type: string, enum: [pending, repo_created, seeding, ready, error] }
                fcStatus: { $ref: "#/components/schemas/AppFcStatus" }
                deployError:
                  type: string
                  description: "Reason stored alongside `fcStatus: deploy_error`."
                authMode:
                  $ref: "#/components/schemas/AppAuthMode"
                  description: >-
                    How the deployed app authenticates end users. Side effects
                    (GoTrue OAuth client registration, secrets) are applied in
                    Task 7; the field is accepted here for PATCH updates.
      responses:
        "200":
          description: The updated app.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteApp
      summary: Delete an app (admin required)
      description: |
        Requires `admin` permission (the creator is treated as admin). Tears down
        the FC function, HTTP trigger, OSS artifact, and GoTrue OAuth client;
        revokes Gitea deploy keys and archives the repo under a `deleted-` prefix.
        Postgres schema, local workdirs, and linked sessions (app_id nulled) are
        kept. The orphaned workspace row is archived with the archived repo URL.
      tags: [Apps]
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: App deleted.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean, enum: [true] }
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    RequestId:
      schema:
        type: string
  parameters:
...
    AppId:
      name: appId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    AppType:
      type: string
      description: |
        What kind of app this is. All three deploy through the same pipeline;
        the type selects the starter template the local daemon writes and
        decides whether the deploy provisions a Postgres schema (`data_app`
        only). `fullstack_tanstack_postgres` is the legacy id for `data_app`
        and is still stored on apps created before the split.
      enum: [static_web, slides, data_app, fullstack_tanstack_postgres]
    AppFcStatus:
      type: string
      description: |
        Deploy lifecycle, orthogonal to `provisionStatus` (which tracks the repo
        and its seeding). `null` means the app has never been deployed.
      enum: [not_deployed, awaiting_build, building, deploying, live, deploy_error]
    AppRuntime:
      type: string
      description: |
        FC runtime for the deployed artifact. Phase 1 supports `node` only;
        `container` is reserved for bring-your-own-image deploys.
      enum: [node, container]
      default: node
    AppAuthMode:
      type: string
      description: |
        How the deployed app authenticates end users. `none` means the live URL
        is reachable without login; `platform` uses TeamClu SSO (GoTrue OAuth
        client per app); `third` uses the app's own OAuth client id stored on
        this row (secret in `app_secrets`, never returned here).
      enum: [none, platform, third]
      default: none
    App:
      type: object
      required: [id, teamId, name, slug, type, visibility, provisionStatus, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        teamId: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        type:
          $ref: "#/components/schemas/AppType"
        visibility: { type: string, enum: [personal, team] }
        workspaceId:
          type: [string, "null"]
          format: uuid
          description: >-
            The app's own workspace, created with the app. Its `path` is filled
            in by the desktop once the local daemon reports where it put the
            checkout (`teams/<teamId>/apps/<appId>`); until then the daemon has
            no directory to resolve for this app.
        gitRemoteUrl:
          type: [string, "null"]
          description: >-
            The app's git remote: the Gitea repo provisioned for it (an SSH URL
            — the deploy key we issue cannot authenticate over HTTPS), or the
            repository it was imported from. Set at creation and not editable.
        gitAuthKind:
          type: [string, "null"]
          enum: [gitea_deploy_key, null]
          description: >-
            `gitea_deploy_key` when this deployment provisioned the app's repo
            and holds a deploy key for it — `/git-head` and `/git-credential`
            answer only for those. Null for an app imported from a remote we
            have no credential for; those deploy the local checkout instead.
        gitCommitSha:
          type: [string, "null"]
          pattern: "^[0-9a-f]{7,40}$"
          description: >-
            HEAD commit SHA on the app's Gitea repo at last successful deploy;
            null before the first deploy completes.
        runtime:
          $ref: "#/components/schemas/AppRuntime"
        authMode:
          $ref: "#/components/schemas/AppAuthMode"
        authModePendingRedeploy:
          type: boolean
          description: >-
            `authMode` was changed after the app went live, so the running
            function still enforces the mode it was deployed with — the OAuth
            env is injected at finalize, not on the PATCH. True until the next
            successful deploy. Clients MUST surface this: an app switched to
            `platform` reads as protected while its site is still public.
            Server-derived from `fc_status` and the deployed mode; false for an
            app that has never been live.
        oauthClientId:
          type: [string, "null"]
          description: >-
            Public OAuth client id when `authMode` is `third`, or the GoTrue
            client id when `authMode` is `platform`. Never includes the matching
            secret.
        provisionStatus:
          type: string
          enum: [pending, repo_created, seeding, ready, error]
        fcStatus:
          oneOf:
            - $ref: "#/components/schemas/AppFcStatus"
            - type: "null"
        fcEndpoint:
          type: [string, "null"]
          description: >-
            Function Compute trigger URL of the deployed app; set once
            `fcStatus` is `live`. Its hostname carries a random suffix, so it
            cannot be derived from the app — prefer `publicUrl` when present.
        publicUrl:
          type: [string, "null"]
          description: >-
            Vanity URL (`https://<slug>-<first 8 of id>.<apps domain>`) served
            by the deployment's own reverse proxy. Derived, never stored, and
            null when the deployment has no apps domain configured — then
            `fcEndpoint` is the only address the app has.
        fcFunctionName: { type: [string, "null"] }
        fcRegion: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - missing_auth
                - invalid_json
                - validation_failed
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - upstream_unavailable
                - internal
            message:
              type: string
            requestId:
              type: string
...